Move GameSetup page to appropriate feature folder
